Fix theme/accent not persisting after app restart
app.exit(0) in window-all-closed killed the process before Electron
flushed pending localStorage writes to disk, and before-quit's
win.destroy() skipped renderer teardown too. Removed both so the
normal app.quit() sequence can flush storage before exit.

Also consolidated a duplicated accent-color implementation:
TopBar.vue and Setting.vue each held independent state with
mismatched fallback colors, and neither applied the saved accent
before first paint. New accent.js store (same pattern as theme.js)
is now the single source of truth, applied pre-mount in renderer.js.

Resolves #48
